Implement skeleton app
This commit implements the majority of the fundamental application.
A tree structure is implemented, with an API that returns either an
iterated tree or a comparison to ask the user.
The tree structure is used in a basic page, with a reducer for state,
prompting a user to sorts elements until a tree is complete.
Currently the list is hard-coded.