Update how we track inserted elements
Prior to the change we removed elements from "to be inserted" when we
_first_ attempted to insert, rather than upon successfully inserting.
Furthermore we lose track of all the elements that have been added to
the tree. Switching implementation to this one here will enable features
like "undo" by keeping track of all the elements, regardless of whether
they've been inserted or not.